scripts/requirements.txt
faker
reportlab
python-docx
